Files Uploaded
RandomizerSolo is simply the other 2 files combined into one for use on a solo run - so you don't have to see all the difficulties and mods get re-calculated for each additional teammate. For multi-player runs, i suggest running the "randomizer" file first to get your difficulty and challenge modifiers, and then running the "weapon" file once for each player to get the weapon that each player will be using.